Improve name filter in MitgliedskontoAuswahlDialog
Now in case the text search input is used, we calculate
an "equality" score for each Mitgliedskonto (aka
Mitgliedskontobuchung) entry. The better the match is
the higher the equality score will be. Only the entries
with score == maxScore will be shown.

Before this commit the search input "Hans-Jürgen Himmelblau"
was also matching for "Hans Meyer" even if the better
match "Hans Himmelblau" also existed. With this commit
they would get the scores "Hans Meyer" -> 1x and
"Hans Himmelblau" -> 2x and so only "Hans Himmelblau" will
be shown.